LeetCode: Is the Max elements index after the index of the smallest element
I hade to write a function that can find the smallest and most abundant element in a list. The most significant value should have an index more significant than the lowest index which if true one will subtract the most considerable amount with the smallest value and print the answer if this doesn't happen print zero.