Bubble-sort that counts swaps
Given an array of integers, sort the array in ascending order using the Bubble Sort algorithm above. Once sorted, print the following three lines:

Array is sorted in numSwaps swaps., where swap is the number of swaps that took place.
First Element: firstElement, where FirstElement  is the first element in the sorted array.
Last Element: lastElement, where LastElement is the last element in the sorted array.